refactor(tests): split IgSender tests into 4 focused files
Mirror the Facebook sender test structure by splitting the monolithic
IgSenderTests.cs into:
- IgSenderTests.cs          → constructor guard clauses, Platform, MessageMaxLength, ISender contract
- IgSenderImageFlowTests.cs → NormalizeImage variants + full image upload/publish flow
- IgSenderSendAsyncTests.cs → SendAsync edge cases (null post, no image, empty image, caption truncation)
- IgSenderResilienceTests.cs → transient failure scenarios (blob upload fail/cancel/NotImplemented, HttpRequestException, no-image guard)

Closes #221
